+++ app/src/main/java/com/teamwss/websoso/ui/main/myPage/myActivity/MyActivityFragment.kt
@@ -1,4 +1,4 @@
-package com.teamwss.websoso.ui.myPage.myActivity
+package com.teamwss.websoso.ui.main.myPage.myActivity
 
 import android.os.Bundle
 import android.view.View
@@ -7,7 +7,7 @@ import com.teamwss.websoso.R
 import com.teamwss.websoso.common.ui.base.BaseFragment
 import com.teamwss.websoso.databinding.FragmentMyActivityBinding
 import com.teamwss.websoso.ui.myActivityDetail.MyActivityDetailActivity
-import com.teamwss.websoso.ui.myPage.myActivity.adapter.MyActivityAdapter
+import com.teamwss.websoso.ui.main.myPage.myActivity.adapter.MyActivityAdapter
 import dagger.hilt.android.AndroidEntryPoint
 
 @AndroidEntryPoint
