+++ app/src/main/java/com/teamwss/websoso/ui/main/myPage/myActivity/adapter/MyActivityViewHolder.kt
@@ -1,10 +1,10 @@
-package com.teamwss.websoso.ui.myPage.myActivity.adapter
+package com.teamwss.websoso.ui.main.myPage.myActivity.adapter
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
 import com.teamwss.websoso.databinding.ItemMyActivityBinding
-import com.teamwss.websoso.ui.myPage.myActivity.model.ActivityModel
+import com.teamwss.websoso.ui.main.myPage.myActivity.model.ActivityModel
 
 class MyActivityViewHolder(private val binding: ItemMyActivityBinding) :
     RecyclerView.ViewHolder(binding.root) {
